:meth:`U.provide`: dtype propagation
The resulting ``undarray`` now has the same dtype as the uncertainty
provided.  Before, the resulting ``undarray`` always had a float dtype.

The divison test [test/ad-hoc/2018-04-21_1940.py] still passes, even
though the main ``ua`` variable now turns out as integer.

+++ upy2/core.py
@@ -146,7 +146,7 @@ class U(upy2.context.Protocol):
 
         stddev = numpy.asarray(error) / self.stddevs
         shape = stddev.shape
-        result = undarray(shape=shape)
+        result = undarray(shape=shape, dtype=stddev.dtype)
 
         dependency = upy2.dependency.Dependency(
                 names=upy2.id_generator.get_idarray(shape=shape),
